FEAT: added tables namespace & basic operations

+++ Src/memory/include/memory.hpp
@@ -14,7 +14,7 @@ uint8_t get_id(void);
 
 uint16_t read_single(uint8_t, uint8_t, uint8_t);
 
-void read_multiple(uint8_t, uint8_t, uint8_t, uint8_t *, uint8_t);
+void read_multiple(uint8_t, uint8_t, uint8_t, uint8_t *, uint16_t);
 
 void write_single(uint8_t, uint8_t, uint8_t, uint8_t);
 void write_multiple(uint8_t, uint8_t, uint8_t, uint8_t *, uint8_t);
